pyproject.toml
[project]
name = "hryaki"
version = "0.0.1"
description = "discord bot for hryaks"
authors = [
    {name = "Максим daseoffc@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.13.2"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pynacl (>=1.5.0,<2.0.0)",
    "fastapi[uvicorn] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
